Modify generate() and create getControlBPoint()
Added getControlBPoint() to create a Vector3d object that represents a
point that will assist with creating the Bezier curve. 

Added generate() which calls getControlBPoint(). Creates a hundred
points for each segment and adds them to the generatedPoints Vector3d
list to be returned. After the last iteration it will create an
additional curve from the final control point to the initial control
point to close the curve.