Refactor generate() and create genCurve()
The function generate() will now call genCurve() to get the points of
each segment of the contour between two points. It will still be used to
generate the contour the initial time and as such create the initial
tension points for the contour.

The function genCurve() will take two control points as Vector3d objects
and generate a list of Vector3d objects that represent the segment
curve. It will return this list.