Modify addControlPoint() in Contour.java
The function addControlPoint() now calculates tension points when a new
point is added. When a new point is added it will create two segments
each with their own tension point. The first will have its value stored
within the previous control point so it's tensionX and tensionY are now
updated. The second segment is stored in the new control point that was
added. 

The function also checks to see if the new point is sorted into the
final or initial point of the list of control point. If it is the
function will properly handle updating the previous control point's
tension values and calculating new tension points without throwing out
of bounds exceptions.