Modify deleteContourPoint() in Contour.java
Added on to deleteContourPoint() so that it will recalculate the tension
point values of the control point previous to the one deleted. Also
added in conditional checks to make sure it can correctly handle when
the initial and final points in the list are removed, and when the list
is empty after the final point has been removed.