Add contour control point type integer values
For each contour type, there needs to be a section in the save file for
the control points (used by the user to edit the curve) and the
generated points (used by Dr. Denney's Matlab program).  By putting the
control points in their own section, we can load these control points
(instead of the generated points) for editing by the user.

This commit adds the integer values for the contour control point types
to provide this functionality.