Attempted to implement Annotation Control: Back Burner
ran into errors trying to update the bottom right panel to have a
list of annotations for the current image. revalidate and repaint
were not working. I even tried calling the panel from another class
to update it. FAILED. I can print the list of contours on an image
in the class but it won't update the panel. the Logic is there.
If we have more time I will look at it later.