Modify ImageView, ImageModel, Contour, Vector3d, and ContourCalc
The mouseDragged() function was refactored to mousePushed().

Contour was given findContourPoint() and findTensionPoint() to return
the index value of the point that was near mouse when pushed.

Contour was given moveControlPoint() and moveTensionPoint() to change
the coordinate values. Is called from mouseDragged().

ImageModel had moveTensionPoint() removed because the function didn't do
anything. 

Vector3d was given setX() and setY() methods to change coordinate values
of the control points.

ContourCalc generate() modified to have a conditional check to not undo
the tension point values when a tension point is being dragged.